adding home page when logged in, and allergies page
