+++ demo_client/src/oauth-nhsd-api/Pages/Index.cshtml
@@ -11,6 +11,6 @@
 
 <div class="container">
     <div class="row">
-        <a class="nhsuk-button" id="btnGo" asp-page="/ProtectedPage">Login to demo</a>
+        <a class="nhsuk-button" id="btnGo" asp-page="/Home">Login to demo</a>
     </div>
 </div>
