--- demo_client/src/oauth-nhsd-api/Pages/ProtectedPage.cshtml
@@ -1,22 +1,0 @@
-﻿@page
-@model oauth_nhsd_api.Pages.ProtectedPageModel
-@{
-    ViewData["Title"] = "Protected Page";
-}
-
-<h1>Protected Page</h1>
-
-<div class="container">
-    <div class="row">
-        <div class="col-3">User:</div>
-        <div class="col-9"> @Model.SessionExpires</div>
-    </div>
-    <div class="row">
-        <div class="col-3">Response Information:</div>
-        <div class="col-9">@Model.ResResponse</div>
-    </div>
-    <div class="row">
-        <div class="col-3">Response Content:</div>
-        <div class="col-9">@Model.ResContent</div>
-    </div>
-</div>
